requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
Werkzeug==3.1.3
click==8.3.0
